+++ eras/babbage/impl/test/Test/Cardano/Ledger/Babbage/BinarySpec.hs
@@ -8,4 +8,4 @@ import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Binary as UpgradeSpec
 
 spec :: Spec
-spec = specUpgrade @Babbage @AlonzoTxAuxData True
+spec = specUpgrade @Babbage True