+++ eras/babbage/impl/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.4.3.1
+## 1.4.4.0
 
-*
+* Add `EqRaw` instance for `BabbageTxBody`
 
 ## 1.4.3.0