+++ eras/mary/impl/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version history for `cardano-ledger-mary`
 
-## 1.3.2.1
+## 1.3.3.0
 
-*
+* Add `EqRaw` instance for `MaryTxBody`
 
 ## 1.3.2.0